cmake: don't assume libraries go to `lib/`
The original bugreport suggested to use `${CMAKE_INSTALL_LIBDIR}`,
however, CMake will actually do the right thing by default because it
knows that the target is a library. That's what we're using for
libyang-cpp and sysrepo-cpp, so let's stick with that for uniformity.

Also, remove an obsolete comment about the installation not working
which dates back to the very first commit.

Bug: https://github.com/CESNET/libnetconf2-cpp/issues/2
Change-Id: I978f0c27c37d2610db641067fe775760e0c6b036