CI: TSAN: fall back to older POSIX functions
...because even the latest version of TSAN does not have all the
required interceptors, and therefore all projects which consume
libnetconf2 (or sysrepo, etc) have to be explicitly configured as such.

This has not been caught by the CI because of the way how we're building
these projects. These "new" functions were introduced by an update to
`libnetconf2 which goes via `CzechLight/dependencies`, and which *does*
build the `libnetconf2-cpp`. But that build correctly sets up these
build flags, and it was only this standalone build of `libnetconf2-cpp`
that was missing the required options. (And in addition, this one
doesn't pin `libnetconf2`, but that's by design and not relevant here.)

TL;DR: after upstream `libnetconf2` started using these more advanced
pthread functions, this project requires an immediate update. Fix that
by using the proper build flags.

Change-Id: I37b826b44a2ccb35ccfe1d511066d86b88640542