do not update libnetconf2 to libnetconf2 v3/v4
That version introduced some backward-incompatible API changes in the C
interface (changed logging for example). Also, the YANG models have
changed a lot in a way which totally break on-disk config compatibility,
*and* the support of local users is currently at a very questionable
level.

This has been reported upstream and we have a plan for resolving the
problems, but for now we have to pin that version of libnetconf2 (and
Netopeer2, in the downstream consumer projects).

Change-Id: Id4bf8df378dcf4444f6c0f523e992cb7f432913b