remove support for (direct) SSH connections
We have a single downstream consumer, netconf-cli, and they have
switched away from using libssh's built-in SSH client a long time ago,
and the connections are done over a FD to ssh(1) process. I'm porting to
libyang-v3 and to the "new", 2023-era libnetconf2 API in the next
commit, so let's reduce this library to a minimal NETCONF client
wrapper.

Change-Id: Id430115e06c08c7de1ed3bf100bdcf03ce1a9321