port to libnetconf2 v3
This is really just a change of logging, nothing else. Unfortunately,
the new logging still affects a global state, it was really just a
change of the function signature, and the logger callback might well be
invoked with a NULL session. This means that mapping from the
nc_session* to a libnetconf::client::Session is not possible, so I
though that simply providing an opaque pointer is not a bad solution.

Change-Id: Iaa10855f4173fdd5e3046efa1827e4931cd05acb
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/libyang-cpp/+/7049