fix parsing rpc-reply
Upstream commit 01a2614 [1] changed the structure of rpc-error subtree,
some of the nodes stopped being opaque. This broke netconf-cli (see
netconf-cli#25) [2].

it is interesting that some of the nodes created by libnetconf2 are
opaque and some are not, but maintainer confirmed that this is OK.
I decided to alter the pattern for checking if the node is error-message
or error-path by checking for the node's path to avoid
node->isOpaque && node->asOpaque().name().name checks. This way it is
also clear we are looking for rpc-error.

[1] https://github.com/CESNET/libnetconf2/commit/01a26147d4dec94a748cfb5fbb6e9fd4f3da1ff6
[2] https://github.com/CESNET/netconf-cli/issues/25

Bug: https://github.com/CESNET/netconf-cli/issues/25
Change-Id: Icf1510d0c591218873cf579fd1de191da35eefa3