another fix at parsing error-path node
The previous commit fe77874cb ("fix parsing rpc-reply") was wrong.
error-path node is *not* opaque as was stated by the commit's message
and it can be seen in the linked upstream commit [1].

The tests worked, because the error-path content was set to a path in an
unimplemented YANG model. The node was then parsed as opaque.

[1] https://github.com/CESNET/libnetconf2/blob/01a26147d4dec94a748cfb5fbb6e9fd4f3da1ff6/src/messages_server.c#L587

Fixes: fe77874cba9b046f5be6aca4727a8c32796ff4e0 ("fix parsing rpc-reply")
Change-Id: I80dbe4b9f544022ae108060e2f5d9fd06edef3f6